+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
   - For successful operations, retry stats can be retrieved using Result.getRetryStats().
   - Otherwise, the original Request may have retry stats available via Request.getRetryStats() (for example, after an exception was thrown).
 - Cloud only: New regions: ap-chiyoda-1, me-dubai-1, uk-cardiff-1 and sa-santiago-1
+- Cloud only: Added new SignatureProvider constructors to allow use of an instance
+principal with a delegation token for authorization and authentication:
+ - SignatureProvider.createInstancePrincipalForDelegation()
 
 ### Fixed
 - Ensure that TableLimits is always null in TableResult on-premise.